Add Select Sync Document Menu (#19)
* Add Page and Screen Select Dialog

* Add Select Sync Document Menu

* fix miss

* method to determine if a node is a screen or not

* delete test code

* Bug Fix: No download if screen is specified but no page is specified

* change modal open mode

* refactoring: correction of variable names, etc

* to return the node instead of the name of the node

* Add Editor class for UnityFigmaBridgeSettings

* delete Dialog

* Added Only Import Selected Pages item

* response to review remarks

* fix error dialog messages

* Add object cleanup

* refactoring

* fix: The process is incorrect due to the order of the ids

* When an error occurs due to an update of the Figma site, call the configuration file

+++ UnityFigmaBridge/Editor/Utils/FigmaPaths.cs
@@ -107,7 +107,7 @@ namespace UnityFigmaBridge.Editor.Utils
             return System.Text.RegularExpressions.Regex.Replace(name, invalidRegStr, "_");
         }
 
-        public static void CreateRequiredDirectories()
+        public static void CreateRequiredDirectories(List<Node> downloadPageNodeList, List<Node> downloadScreenNodeList)
         {
             
             //  Create directory for pages if required 
@@ -115,9 +115,12 @@ namespace UnityFigmaBridge.Editor.Utils
             {
                 Directory.CreateDirectory(FigmaPagePrefabFolder);
             }
+
             // Remove existing prefabs for pagwes
+            var downloadPageNameList = downloadPageNodeList.Select(p => p.name).ToList();
             foreach (var file in new DirectoryInfo(FigmaPagePrefabFolder).GetFiles())
             {
+                if (!downloadPageNameList.Contains(file.Name)) continue;
                 file.Delete(); 
             }
             
@@ -127,8 +130,10 @@ namespace UnityFigmaBridge.Editor.Utils
                 Directory.CreateDirectory(FigmaScreenPrefabFolder);
             }
             // Remove existing flowScreen prefabs
+            var downloadScreenNameList = downloadScreenNodeList.Select(s => s.name).ToList();
             foreach (FileInfo file in  new DirectoryInfo(FigmaScreenPrefabFolder).GetFiles())
             {
+                if (!downloadScreenNameList.Contains(file.Name)) continue;
                 file.Delete(); 
             }
